+++ src/main/java/org/enigma/service/ProductService.java
@@ -1,21 +1,21 @@
 package org.enigma.service;
 
-import org.enigma.model.Store;
-import org.enigma.repository.StoreRepository;
+import org.enigma.model.Product;
+import org.enigma.repository.ProductRepository;
 import java.util.List;
 import java.util.Optional;
 
-public class StoreService implements IStoreService {
-    private final StoreRepository repository;
+public class ProductService implements IStoreService {
+    private final ProductRepository repository;
 
-    public StoreService(StoreRepository repository) {
+    public ProductService(ProductRepository repository) {
         this.repository = repository;
     }
 
     @Override
-    public List<Store> list() {
+    public List<Product> list() {
         try {
-            List<Store> list = repository.getAll();
+            List<Product> list = repository.getAll();
             if(list.isEmpty()){
                 System.out.println("Data is empty");
             }
@@ -26,17 +26,17 @@ public class StoreService implements IStoreService {
     }
 
     @Override
-    public Store create(Store store) {
+    public Product create(Product product) {
         try {
             System.out.println("Product Added");
-            return repository.create(store);
+            return repository.create(product);
         }catch (Exception e){
             throw new RuntimeException(e.getMessage());
         }
     }
 
     @Override
-    public void update(Store update, String id) {
+    public void update(Product update, String id) {
         try {
             System.out.println("Product Updated");
             repository.update(update,id);
@@ -48,15 +48,20 @@ public class StoreService implements IStoreService {
     @Override
     public void delete(String id) {
         try {
-            System.out.println("Product Deleted");
-            repository.delete(id);
+            Optional<Product> find = repository.findById(id);
+            if(find.isEmpty()){
+                System.out.println("Id not found");
+            }else {
+                System.out.println("Product Deleted");
+                repository.delete(id);
+            }
         }catch (Exception e){
             throw new RuntimeException(e.getMessage());
         }
     }
 
     @Override
-    public Optional<Store> findId(String id) {
+    public Optional<Product> findId(String id) {
         try {
             return repository.findById(id);
         }catch (Exception e){
@@ -65,7 +70,7 @@ public class StoreService implements IStoreService {
     }
 
     @Override
-    public List<Store> findByProduct(String product) {
+    public List<Product> findByProduct(String product) {
         try {
             return repository.findByProduct(product);
         }catch (Exception e){
@@ -74,7 +79,7 @@ public class StoreService implements IStoreService {
     }
 
     @Override
-    public List<Store> findBySize(String size) {
+    public List<Product> findBySize(String size) {
         try {
             return repository.findBySize(size);
         }catch (Exception e){
